Split the Locking of the Pool
In order to support scalable concurrent insertions by different threads
a new property org.eclipse.emf.common.util.Pool.concurrencyLevel is
introduced. The property defaults to 1, which is the concurrency level
that was supported until now.

When higher concurrency levels are set, the pool is made thread-safe
using as many locks as the concurrency level. To decide which lock is
take for each entry, the hashCode modulo concurrencyLevel is used. Thus,
different threads will usually insert entries in the pool without
blocking each other as long as the hashCodes are well distributed.

The downside is that locking the collection for exclusive access is more
costly, as it needs to acquire all locks in sequence.